ISSUE #5982 - Moved groupbyHelper to a more suitable directory
