+++ frontend/src/v5/ui/components/tickets/ticketsGroupBy.helper.ts
